feat:porting to npx
