+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "custom-react-hooks",
-  "displayName": "React Hooks",
+  "name": "common-api-hooks",
+  "displayName": "Custom React Hooks",
   "description": "A collection of customizable React hooks to accelerate your workflow.",
   "version": "0.0.1",
   "publisher": "AyushmaanSingh",
@@ -22,43 +22,17 @@
     "Custom Hooks",
     "Custom React Hooks",
     "Productivity",
-    "VS Code Extension"
-],
-  "activationEvents": [
-    "onCommand:react-hooks.createHooksFolder"
+    "NPX"
   ],
-  "main": "./dist/extension.js",
-  "contributes": {
-    "commands": [
-      {
-        "command": "react-hooks.createHooksFolder",
-        "title": "React Hooks"
-      }
-    ]
-  },
   "scripts": {
-    "vscode:prepublish": "npm run package",
-    "compile": "webpack",
-    "watch": "webpack --watch",
-    "package": "webpack --mode production --devtool hidden-source-map",
-    "compile-tests": "tsc -p . --outDir out",
-    "watch-tests": "tsc -p . -w --outDir out",
-    "pretest": "npm run compile-tests && npm run compile && npm run lint",
-    "lint": "eslint src",
-    "test": "vscode-test"
+    "build": "tsc",
+    "start": "node dist/create-hooks.js"
   },
   "devDependencies": {
-    "@types/vscode": "^1.94.0",
-    "@types/mocha": "^10.0.8",
     "@types/node": "20.x",
-    "@typescript-eslint/eslint-plugin": "^8.7.0",
-    "@typescript-eslint/parser": "^8.7.0",
-    "eslint": "^9.11.1",
-    "typescript": "^5.6.2",
-    "ts-loader": "^9.5.1",
-    "webpack": "^5.95.0",
-    "webpack-cli": "^5.1.4",
-    "@vscode/test-cli": "^0.0.10",
-    "@vscode/test-electron": "^2.4.1"
+    "typescript": "^5.6.2"
+  },
+  "bin": {
+    "create-react-hooks": "./dist/create-hooks.js"
   }
 }
